requirements.txt
blinker==1.9.0
click==8.3.3
contourpy==1.3.3
cycler==0.12.1
filelock==3.29.0
Flask==3.1.3
fonttools==4.62.1
fsspec==2026.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.2
pandas==3.0.2
pillow==12.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==81.0.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.11.0
torchvision==0.26.0
typing_extensions==4.15.0
Werkzeug==3.1.8
